REFACTOR: Instanciate components folder
